+++ slangpy/tests/device/test_bindless.py
@@ -15,8 +15,6 @@ def test_bindless_texture(device_type: spy.DeviceType):
     device = helpers.get_device(device_type)
     if not device.has_feature(spy.Feature.bindless):
         pytest.skip("Bindless not supported on this device.")
-    if device_type == spy.DeviceType.vulkan:
-        pytest.skip("Due to a slang bug bindless samplers are currently not supported on Vulkan.")
 
     module = device.load_module("test_bindless_texture.slang")
     program = device.link_program(