+++ docs/changelog.rst
@@ -10,9 +10,10 @@ SlangPy uses a `semantic versioning <http://semver.org>`__ policy for its API.
 Version 0.31.0 (TBA)
 ----------------------------
 
-- Add support for vectorizing against Python lists
-- `NDBuffer` and `Tensor` `empty` / `zeros` apis consistent
-- Added `load_from_image` for `NDBuffer` and `Tensor`
+- Update to slang version 2025.10.1.
+- Add support for vectorizing against Python lists.
+- Make ``NDBuffer`` and ``Tensor`` ``empty`` / ``zeros`` APIs consistent.
+- Added ``load_from_image`` for ``NDBuffer`` and ``Tensor``.
 - Fix typings for ``float2x3``, ``float3x2``, ``float4x2`` and ``float4x3``.
 
 Version 0.30.0 (May 27, 2025)